% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_mse}
\alias{calculate_mse}
\title{Calculate the Mean Squared Error (MSE)}
\usage{
calculate_mse(y_true, y_pred)
}
\arguments{
\item{y_true}{The vector of true values.}

\item{y_pred}{The vector of predicted values.}
}
\value{
Returns the Mean Squared Error (MSE).
}
\description{
Calculate the Mean Squared Error (MSE)
}
