% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_r2}
\alias{calculate_r2}
\title{Calculate the R² value}
\usage{
calculate_r2(y_true, y_pred)
}
\arguments{
\item{y_true}{The vector of true values.}

\item{y_pred}{The vector of predicted values.}
}
\value{
Returns the R² value, which represents the goodness of fit of the model.
}
\description{
Calculate the R² value
}
