% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_regression}
\alias{linear_regression}
\title{Fit a Linear Regression Model using lm()}
\usage{
linear_regression(data, formula)
}
\arguments{
\item{data}{The input data.}

\item{formula}{The formula for the regression model.}
}
\value{
The results of the fitted regression model.
}
\description{
Fit a Linear Regression Model using lm()
}
\examples{
data <- data.frame(x = c(1, 2, 3, 4), y = c(2, 4, 6, 8))
formula <- y ~ x
linear_regression(data, formula)
}
