% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_features.R
\name{select_features}
\alias{select_features}
\title{Feature Selection}
\usage{
select_features(X, y, threshold = 0.2)
}
\arguments{
\item{X}{The input data matrix, where each row represents a sample and each column represents a feature.}

\item{y}{The target variable vector.}

\item{threshold}{The correlation threshold used to select features correlated with the target variable, default is 0.2.}
}
\value{
The selected feature matrix.
}
\description{
Feature Selection
}
\examples{
X <- matrix(c(1, 2, 3, 4, 5, 6), ncol = 2)
y <- c(1, 2, 3)
select_features(X, y)
}
